Added Health and Double Jump Items
Items spawn in the display area and disappear when touched by the player. Health increases health of player and is indicated in the display. Double Jump allows player to jump another time while airborne. This pickup is also shown in the display and changes color of character while on a platform.